Move forHourOfDay method inside the TimeOfDay enum
