+++ FacesofLondon/app/src/main/java/com/novoda/wallpaper/droidcon/TimeOfDay.java
@@ -1,5 +1,7 @@
 package com.novoda.wallpaper.droidcon;
 
+import java.util.Locale;
+
 enum TimeOfDay {
     DAWN(4),
     DAY(10),
@@ -12,7 +14,22 @@ enum TimeOfDay {
         this.hourOfDay = hourOfDay;
     }
 
-    int hourOfDay() {
-        return hourOfDay;
+    public static TimeOfDay forHourOfDay(int hourOfDay) {
+        mustBeValidHourOfDay(hourOfDay);
+        if (hourOfDay >= DAWN.hourOfDay && hourOfDay < DAY.hourOfDay) {
+            return DAWN;
+        } else if (hourOfDay >= DAY.hourOfDay && hourOfDay < DUSK.hourOfDay) {
+            return DAY;
+        } else if (hourOfDay >= DUSK.hourOfDay && hourOfDay < NIGHT.hourOfDay) {
+            return DUSK;
+        } else {
+            return NIGHT;
+        }
+    }
+
+    private static void mustBeValidHourOfDay(int hourOfDay) {
+        if (hourOfDay < 0 || hourOfDay > 24) {
+            throw new IllegalArgumentException(String.format(Locale.US, "A day only has 24 hours. Got %d instead", hourOfDay));
+        }
     }
 }
